DEI UMA CORREÇÃO NO QUE JOSÉ IMPLEMENTOU PARA DEIXAR O PROJETO PADRONIZADO E DEIXEI O CRUD PRONTO

+++ src/main/java/br/com/dcx/ufpb/eng/ApiCiclismo/service/ProfileServiceImpl.java
@@ -1,20 +1,21 @@
 package br.com.dcx.ufpb.eng.ApiCiclismo.service;
 
+import br.com.dcx.ufpb.eng.ApiCiclismo.dto.ProfileDTO;
 import br.com.dcx.ufpb.eng.ApiCiclismo.entity.Profile;
 import br.com.dcx.ufpb.eng.ApiCiclismo.enums.CyclingCategory;
+import br.com.dcx.ufpb.eng.ApiCiclismo.exception.ProfileNotFoudException;
 import br.com.dcx.ufpb.eng.ApiCiclismo.repositories.ProfileRepository;
 import org.hibernate.usertype.UserType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Service
 public class ProfileServiceImpl implements ProfileService {
 
-
-    // falta implementar
 
     private ProfileRepository profileRepository;
 
@@ -23,16 +24,43 @@ public class ProfileServiceImpl implements ProfileService {
         this.profileRepository = profileRepository;
     }
 
+    public List<Profile> getProfiles() {
+        return profileRepository.findAll();
+    }
+
     public Profile saveProfile(Profile profile) {
         return profileRepository.save(profile);
     }
 
-    public Profile getProfileById(Long id) {
-        return profileRepository.getReferenceById(id);
+    public Optional<Profile> getProfileById(Long id) throws ProfileNotFoudException {
+        Optional<Profile> optionalProfile = profileRepository.findById(id);
+        if (optionalProfile.isEmpty()) {
+            throw new ProfileNotFoudException("Profile not found.");
+        }
+        return optionalProfile;
     }
 
-    public void deleteProfile(Long id) {
+    public void deleteProfile(Long id) throws ProfileNotFoudException {
+        Optional<Profile> optionalProfile = profileRepository.findById(id);
+        if (optionalProfile.isEmpty()) {
+            throw new ProfileNotFoudException("Profile not found.");
+        }
         profileRepository.deleteById(id);
+    }
+
+    public void updateProfile(Long id, ProfileDTO profileDTO) throws ProfileNotFoudException {
+        Optional<Profile> optionalProfile = profileRepository.findById(id);
+        if (optionalProfile.isEmpty()) {
+            throw new ProfileNotFoudException("Profile not found.");
+        }
+        Profile profile = optionalProfile.get();
+        profile.setFullName(profileDTO.getFullName());
+        profile.setNickname(profileDTO.getNickname());
+        profile.setSexo(profileDTO.getSexo());
+        profile.setLocation(profileDTO.getLocation());
+        profile.setCiclistaProfissional();
+        profile.setCiclistaAmador();
+        profileRepository.save(profile);
     }
 
     public List<Profile> getProfilesByFullName(String fullName) {
@@ -55,17 +83,4 @@ public class ProfileServiceImpl implements ProfileService {
     public List<Profile> getProfilesByLocation(String location) {
         return profileRepository.findByLocation(location);
     }
-
-    public Profile updateProfile(Long id, ProfileService updatedProfile) {
-        return null;
-    }
-
-    public Profile readProfile(Long id) {
-        return profileRepository.getReferenceById(id);
-    }
-
-    public Profile createProfile(Profile profile) {
-        return profileRepository.save(profile);
-    }
-
 }
